Refactor LOG imports to resolve errors
